[Refactor] ReplyService Interface
