+++ poorband/src/main/java/com/finalproject/mvc/sobeit/service/ReplyService.java
@@ -1,45 +1,11 @@
 package com.finalproject.mvc.sobeit.service;
 
 import com.finalproject.mvc.sobeit.entity.Reply;
-import com.finalproject.mvc.sobeit.repository.ReplyRepo;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
+public interface ReplyService {
+    public Reply writeReply(Reply reply);
 
-@Service
-@RequiredArgsConstructor
-public class ReplyService {
+    public Reply updateReply(Reply reply);
 
-    private final ReplyRepo replyRepo;
-
-    /**
-     * 댓글 작성
-     * @param reply
-     */
-    public Reply writeReply(Reply reply){
-        reply.setWrittenDate(LocalDateTime.now());
-        return replyRepo.save(reply);
-    }
-
-    /**
-     * 댓글 수정
-     * @param reply
-     * @return
-     */
-    public Reply updateReply(Reply reply){
-        reply.setWrittenDate(LocalDateTime.now());
-        // 수정일 갱신
-        return replyRepo.save(reply);
-    }
-
-    /**
-     * 댓글 삭제
-     * @param id
-     */
-    public void deleteReply(Long id){
-        replyRepo.deleteById(id);
-        // 자식 댓글 다 삭제해야 댐ㅎㅎ
-    }
+    public void deleteReply(Long id);
 }
